First full upload of relevant files.
Most files are to be ran line by line and do not contain full documented functions.

Exception is the NFI_02_TaperVolume_BC.R which contains the structure of the final function but needs to be modified to work with Kozak's table with all parameters. Currently it only has a single example based on a NFI site.